backend/requirements.txt
fastapi
uvicorn
python-jose
httpx
cassandra-driver